タスク8　商品削除機能修正

+++ app/controllers/items_controller.rb
@@ -23,9 +23,9 @@ class ItemsController < ApplicationController
   end
 
   def edit
-    unless current_user.id == @item.user.id
-      redirect_to action: :index 
-    end
+    return if current_user.id == @item.user.id
+
+    redirect_to action: :index
   end
 
   def update
@@ -35,11 +35,15 @@ class ItemsController < ApplicationController
       render :edit
     end
   end
- 
+
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if user_signed_in? && current_user.id = item.user.id
+      item.destroy
+      redirect_to root_path
+    else
+      redirect_to root_path
+    end
   end
 
   private
@@ -49,7 +53,7 @@ class ItemsController < ApplicationController
                                  :day_to_send_id, :price, :image).merge(user_id: current_user.id)
   end
 
-  def set_item 
+  def set_item
     @item = Item.find(params[:id])
   end
 end
